Refactor ItemBase, Item, ItemClone and their tests
